Trait nautilus_infrastructure::cache::CacheDatabase

pub trait CacheDatabase {
    type DatabaseType;

    // Required methods
    fn new(
        trader_id: TraderId,
        instance_id: UUID4,
        config: HashMap<String, Value>
    ) -> Result<Self::DatabaseType>;
    fn flushdb(&mut self) -> Result<()>;
    fn keys(&mut self, pattern: &str) -> Result<Vec<String>>;
    fn read(&mut self, key: &str) -> Result<Vec<Vec<u8>>>;
    fn insert(
        &mut self,
        key: String,
        payload: Option<Vec<Vec<u8>>>
    ) -> Result<()>;
    fn update(
        &mut self,
        key: String,
        payload: Option<Vec<Vec<u8>>>
    ) -> Result<()>;
    fn delete(
        &mut self,
        key: String,
        payload: Option<Vec<Vec<u8>>>
    ) -> Result<()>;
    fn handle_messages(
        rx: Receiver<DatabaseCommand>,
        trader_key: String,
        config: HashMap<String, Value>
    );
}

Provides a generic cache database facade.

The main operations take a consistent key and payload which should provide enough information to implement the cache database in many different technologies.

Delete operations may need a payload to target specific values.

Required Associated Types

type DatabaseType

Required Methods

fn new( trader_id: TraderId, instance_id: UUID4, config: HashMap<String, Value> ) -> Result<Self::DatabaseType>

fn flushdb(&mut self) -> Result<()>

fn keys(&mut self, pattern: &str) -> Result<Vec<String>>

fn read(&mut self, key: &str) -> Result<Vec<Vec<u8>>>

fn insert(&mut self, key: String, payload: Option<Vec<Vec<u8>>>) -> Result<()>

fn update(&mut self, key: String, payload: Option<Vec<Vec<u8>>>) -> Result<()>

fn delete(&mut self, key: String, payload: Option<Vec<Vec<u8>>>) -> Result<()>

fn handle_messages( rx: Receiver<DatabaseCommand>, trader_key: String, config: HashMap<String, Value> )

Object Safety

This trait is not object safe.