nautilus_infrastructure::redis::msgbus

Struct RedisMessageBusDatabase

pub struct RedisMessageBusDatabase {
    pub trader_id: TraderId,
    pub instance_id: UUID4,
    /* private fields */
}

Fields
trader_id: TraderId

The trader ID for this message bus database.

instance_id: UUID4

The instance ID for this message bus database.

Implementations

impl RedisMessageBusDatabase

pub fn get_stream_receiver(&mut self) -> Result<Receiver<BusMessage>>

Retrieves the Redis stream receiver for this message bus instance.

Errors

Returns an error if the stream receiver has already been taken.

pub fn stream( stream_rx: Receiver<BusMessage>, ) -> impl Stream<Item = BusMessage> + 'static

Streams messages arriving on the stream receiver channel.

pub async fn close_async(&mut self)