nautilus_infrastructure::sql::cache

Struct PostgresCacheDatabase

pub struct PostgresCacheDatabase {
    pub pool: PgPool,
    /* private fields */
}

Fields
pool: PgPool
Implementations

impl PostgresCacheDatabase

pub async fn connect( host: Option<String>, port: Option<u16>, username: Option<String>, password: Option<String>, database: Option<String>, ) -> Result<Self, Error>

Connects to the Postgres cache database using the provided connection parameters.

Errors

Returns an error if establishing the database connection fails.

Panics

Panics if the internal Postgres pool connection attempt (connect_pg) unwraps on error.